Basic Syntax...First assignment

+++ BasicSyntax.py/Temperature.py
@@ -1,3 +1,3 @@
-celcuis= float(input("Enter the temperature in celcuis: "))
-fahrenheit=float(celcuis*9/5+32)
+celcuis= int(input("Enter the temperature in celcuis: "))
+fahrenheit=(celcuis*9/5+32)
 print ("Coverted", celcuis, "celcuis into fahrenheit =", fahrenheit, "fahrenheit" )
